+++ packages/amplify-category-auth/src/provider-utils/awscloudformation/utils/attach-prev-params-to-context.ts
@@ -1,11 +1,11 @@
 import { getAuthResourceName } from '../../../utils/getAuthResourceName';
-import { supportedServices } from '../../supported-services';
+import { getSupportedServices } from '../../supported-services';
 
 /**
  * Some existing code relies on the existing auth parameters being attached to context.updatingAuth
  */
 export const attachPrevParamsToContext = async (context: any) => {
   const resourceName = await getAuthResourceName(context);
-  const providerPlugin = context.amplify.getPluginInstance(context, supportedServices.Cognito.provider);
+  const providerPlugin = context.amplify.getPluginInstance(context, getSupportedServices().Cognito.provider);
   context.updatingAuth = providerPlugin.loadResourceParameters(context, 'auth', resourceName);
 };