+++ packages/graphql-transformers-e2e-tests/src/__tests__/PerFieldAuthTests.e2e.test.ts
@@ -16,10 +16,11 @@ import { default as moment } from 'moment';
 import {
   createUserPool,
   createUserPoolClient,
-  signupAndAuthenticateUser,
   createGroup,
   addUserToGroup,
   configureAmplify,
+  signupUser,
+  authenticateUser,
 } from '../cognitoUtils';
 import 'isomorphic-fetch';
 
@@ -220,8 +221,9 @@ beforeAll(async () => {
     // Configure Amplify, create users, and sign in.
     configureAmplify(USER_POOL_ID, userPoolClientId);
 
-    await signupAndAuthenticateUser(USER_POOL_ID, USERNAME1, TMP_PASSWORD, REAL_PASSWORD);
-    await signupAndAuthenticateUser(USER_POOL_ID, USERNAME2, TMP_PASSWORD, REAL_PASSWORD);
+    await signupUser(USER_POOL_ID, USERNAME1, TMP_PASSWORD);
+    await signupUser(USER_POOL_ID, USERNAME2, TMP_PASSWORD);
+    await signupUser(USER_POOL_ID, USERNAME3, TMP_PASSWORD);
     await createGroup(USER_POOL_ID, ADMIN_GROUP_NAME);
     await createGroup(USER_POOL_ID, PARTICIPANT_GROUP_NAME);
     await createGroup(USER_POOL_ID, WATCHER_GROUP_NAME);
@@ -234,15 +236,15 @@ beforeAll(async () => {
     await addUserToGroup(INSTRUCTOR_GROUP_NAME, USERNAME1, USER_POOL_ID);
     await addUserToGroup(INSTRUCTOR_GROUP_NAME, USERNAME2, USER_POOL_ID);
 
-    const authResAfterGroup: any = await signupAndAuthenticateUser(USER_POOL_ID, USERNAME1, TMP_PASSWORD, REAL_PASSWORD);
+    const authResAfterGroup: any = await authenticateUser(USERNAME1, TMP_PASSWORD, REAL_PASSWORD);
     const idToken = authResAfterGroup.getIdToken().getJwtToken();
     GRAPHQL_CLIENT_1 = new GraphQLClient(GRAPHQL_ENDPOINT, { Authorization: idToken });
 
-    const authRes2AfterGroup: any = await signupAndAuthenticateUser(USER_POOL_ID, USERNAME2, TMP_PASSWORD, REAL_PASSWORD);
+    const authRes2AfterGroup: any = await authenticateUser(USERNAME2, TMP_PASSWORD, REAL_PASSWORD);
     const idToken2 = authRes2AfterGroup.getIdToken().getJwtToken();
     GRAPHQL_CLIENT_2 = new GraphQLClient(GRAPHQL_ENDPOINT, { Authorization: idToken2 });
 
-    const authRes3: any = await signupAndAuthenticateUser(USER_POOL_ID, USERNAME3, TMP_PASSWORD, REAL_PASSWORD);
+    const authRes3: any = await authenticateUser(USERNAME3, TMP_PASSWORD, REAL_PASSWORD);
     const idToken3 = authRes3.getIdToken().getJwtToken();
     GRAPHQL_CLIENT_3 = new GraphQLClient(GRAPHQL_ENDPOINT, { Authorization: idToken3 });