+++ packages/amplify-appsync-simulator/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.27.7](https://github.com/aws-amplify/amplify-cli/compare/amplify-appsync-simulator@1.27.6...amplify-appsync-simulator@1.27.7) (2021-08-24)
+
+**Note:** Version bump only for package amplify-appsync-simulator
+
+
+
+
+
 ## [1.27.6](https://github.com/aws-amplify/amplify-cli/compare/amplify-appsync-simulator@1.27.5...amplify-appsync-simulator@1.27.6) (2021-08-06)
 
 **Note:** Version bump only for package amplify-appsync-simulator