+++ packages/amplify-category-analytics/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.21.21](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-analytics@2.21.20...amplify-category-analytics@2.21.21) (2021-09-18)
+
+**Note:** Version bump only for package amplify-category-analytics
+
+
+
+
+
+## [2.21.20](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-analytics@2.21.19...amplify-category-analytics@2.21.20) (2021-09-14)
+
+**Note:** Version bump only for package amplify-category-analytics
+
+
+
+
+
+## [2.21.19](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-analytics@2.21.18...amplify-category-analytics@2.21.19) (2021-09-09)
+
+**Note:** Version bump only for package amplify-category-analytics
+
+
+
+
+
+## [2.21.18](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-analytics@2.21.17...amplify-category-analytics@2.21.18) (2021-09-02)
+
+**Note:** Version bump only for package amplify-category-analytics
+
+
+
+
+
+## [2.21.17](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-analytics@2.21.16...amplify-category-analytics@2.21.17) (2021-08-24)
+
+**Note:** Version bump only for package amplify-category-analytics
+
+
+
+
+
 ## [2.21.16](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-analytics@2.21.15...amplify-category-analytics@2.21.16) (2021-08-06)
 
 **Note:** Version bump only for package amplify-category-analytics