+++ packages/amplify-category-predictions/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.9.12](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-predictions@2.9.11...amplify-category-predictions@2.9.12) (2021-09-18)
+
+**Note:** Version bump only for package amplify-category-predictions
+
+
+
+
+
+## [2.9.11](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-predictions@2.9.10...amplify-category-predictions@2.9.11) (2021-09-14)
+
+**Note:** Version bump only for package amplify-category-predictions
+
+
+
+
+
+## [2.9.10](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-predictions@2.9.9...amplify-category-predictions@2.9.10) (2021-09-09)
+
+**Note:** Version bump only for package amplify-category-predictions
+
+
+
+
+
+## [2.9.9](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-predictions@2.9.8...amplify-category-predictions@2.9.9) (2021-09-02)
+
+**Note:** Version bump only for package amplify-category-predictions
+
+
+
+
+
+## [2.9.8](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-predictions@2.9.7...amplify-category-predictions@2.9.8) (2021-08-24)
+
+**Note:** Version bump only for package amplify-category-predictions
+
+
+
+
+
 ## [2.9.7](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-predictions@2.9.6...amplify-category-predictions@2.9.7) (2021-08-06)
 
 **Note:** Version bump only for package amplify-category-predictions