+++ packages/amplify-frontend-ios/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.20.14](https://github.com/aws-amplify/amplify-cli/compare/amplify-frontend-ios@2.20.13...amplify-frontend-ios@2.20.14) (2021-09-18)
+
+**Note:** Version bump only for package amplify-frontend-ios
+
+
+
+
+
+## [2.20.13](https://github.com/aws-amplify/amplify-cli/compare/amplify-frontend-ios@2.20.12...amplify-frontend-ios@2.20.13) (2021-09-14)
+
+**Note:** Version bump only for package amplify-frontend-ios
+
+
+
+
+
+## [2.20.12](https://github.com/aws-amplify/amplify-cli/compare/amplify-frontend-ios@2.20.11...amplify-frontend-ios@2.20.12) (2021-09-09)
+
+**Note:** Version bump only for package amplify-frontend-ios
+
+
+
+
+
+## [2.20.11](https://github.com/aws-amplify/amplify-cli/compare/amplify-frontend-ios@2.20.10...amplify-frontend-ios@2.20.11) (2021-09-02)
+
+**Note:** Version bump only for package amplify-frontend-ios
+
+
+
+
+
+## [2.20.10](https://github.com/aws-amplify/amplify-cli/compare/amplify-frontend-ios@2.20.9...amplify-frontend-ios@2.20.10) (2021-08-24)
+
+**Note:** Version bump only for package amplify-frontend-ios
+
+
+
+
+
 ## [2.20.9](https://github.com/aws-amplify/amplify-cli/compare/amplify-frontend-ios@2.20.8...amplify-frontend-ios@2.20.9) (2021-08-06)