+++ packages/amplify-prompts/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.2](https://github.com/aws-amplify/amplify-cli/compare/amplify-prompts@1.1.1...amplify-prompts@1.1.2) (2021-09-02)
+
+**Note:** Version bump only for package amplify-prompts
+
+
+
+
+
+## [1.1.1](https://github.com/aws-amplify/amplify-cli/compare/amplify-prompts@1.1.0...amplify-prompts@1.1.1) (2021-08-24)
+
+**Note:** Version bump only for package amplify-prompts
+
+
+
+
+
 # 1.1.0 (2021-08-06)