+++ packages/amplify-util-mock/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.34.5](https://github.com/aws-amplify/amplify-cli/compare/amplify-util-mock@3.34.4...amplify-util-mock@3.34.5) (2021-09-18)
+
+**Note:** Version bump only for package amplify-util-mock
+
+
+
+
+
+## [3.34.4](https://github.com/aws-amplify/amplify-cli/compare/amplify-util-mock@3.34.3...amplify-util-mock@3.34.4) (2021-09-14)
+
+**Note:** Version bump only for package amplify-util-mock
+
+
+
+
+
+## [3.34.3](https://github.com/aws-amplify/amplify-cli/compare/amplify-util-mock@3.34.2...amplify-util-mock@3.34.3) (2021-09-09)
+
+**Note:** Version bump only for package amplify-util-mock
+
+
+
+
+
+## [3.34.2](https://github.com/aws-amplify/amplify-cli/compare/amplify-util-mock@3.34.1...amplify-util-mock@3.34.2) (2021-09-02)
+
+**Note:** Version bump only for package amplify-util-mock
+
+
+
+
+
+## [3.34.1](https://github.com/aws-amplify/amplify-cli/compare/amplify-util-mock@3.34.0...amplify-util-mock@3.34.1) (2021-08-24)
+
+**Note:** Version bump only for package amplify-util-mock
+
+
+
+
+
 # [3.34.0](https://github.com/aws-amplify/amplify-cli/compare/amplify-util-mock@3.33.7...amplify-util-mock@3.34.0) (2021-08-06)