+++ packages/graphql-connection-transformer/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.21.22](https://github.com/aws-amplify/amplify-cli/compare/graphql-connection-transformer@4.21.21...graphql-connection-transformer@4.21.22) (2021-09-18)
+
+**Note:** Version bump only for package graphql-connection-transformer
+
+
+
+
+
+## [4.21.21](https://github.com/aws-amplify/amplify-cli/compare/graphql-connection-transformer@4.21.20...graphql-connection-transformer@4.21.21) (2021-09-14)
+
+**Note:** Version bump only for package graphql-connection-transformer
+
+
+
+
+
+## [4.21.20](https://github.com/aws-amplify/amplify-cli/compare/graphql-connection-transformer@4.21.19...graphql-connection-transformer@4.21.20) (2021-09-09)
+
+**Note:** Version bump only for package graphql-connection-transformer
+
+
+
+
+
+## [4.21.19](https://github.com/aws-amplify/amplify-cli/compare/graphql-connection-transformer@4.21.18...graphql-connection-transformer@4.21.19) (2021-09-02)
+
+**Note:** Version bump only for package graphql-connection-transformer
+
+
+
+
+
+## [4.21.18](https://github.com/aws-amplify/amplify-cli/compare/graphql-connection-transformer@4.21.17...graphql-connection-transformer@4.21.18) (2021-08-24)
+
+**Note:** Version bump only for package graphql-connection-transformer
+
+
+
+
+
 ## [4.21.17](https://github.com/aws-amplify/amplify-cli/compare/graphql-connection-transformer@4.21.16...graphql-connection-transformer@4.21.17) (2021-08-06)
 
 **Note:** Version bump only for package graphql-connection-transformer